177. Add support for building shared libraries on BSD/OS (#3876, Kurt Lidl).  176. Add imake support for automatically determining the version for BSD/OS,       and update bsdi.cf for the features relevant to the different versions       (#3876, Kurt Lidl).  175. Xterm patch #132 (#3874, Thomas Dickey).  174. Fix rman-related build problems when building external applications       (#3869, Matthieu Herrb, #3890, H. J. Lu).  173. Fix a typo in OpenBSD.cf (#3867, Matthieu Herrb).  172. Fix some build warnings (#3864, Ani Joshi).

+++ programs/Xserver/hw/xfree86/loader/elfloader.c
@@ -1,4 +1,4 @@
-/* $XFree86: xc/programs/Xserver/hw/xfree86/loader/elfloader.c,v 1.22 1999/08/22 05:57:38 dawes Exp $ */
+/* $XFree86: xc/programs/Xserver/hw/xfree86/loader/elfloader.c,v 1.23 1999/12/13 23:56:38 robin Exp $ */
 
 /*
  *
@@ -1236,7 +1236,7 @@ ELFDEBUG( "*dest32=%8.8x\n", *dest32 );
 #else
  /* Linux PPC */
 	case R_PPC_ADDR32: /* 1 */
-	    dest32=(unsigned long *)(secp+rel->r_offset);
+	    dest32=(unsigned int *)(secp+rel->r_offset);
 	    symval=ElfGetSymbolValue(elffile,ELF_R_SYM(rel->r_info));
 #ifdef ELFDEBUG
 	    ELFDEBUG( "R_PPC_ADDR32\t" );
@@ -1328,7 +1328,7 @@ ELFDEBUG( "*dest32=%8.8x\n", *dest32 );
 #endif
 	    break;
 	case R_PPC_REL24: /* 10 */
-	    dest32=(unsigned long *)(secp+rel->r_offset);
+	    dest32=(unsigned int *)(secp+rel->r_offset);
 #ifdef ELFDEBUG
 	    ELFDEBUG( "R_PPC_REL24 %s\t", ElfGetSymbolName(elffile,ELF_R_SYM(rel->r_info)) );
 	    ELFDEBUG( "secp=%x\t", secp );
@@ -1370,7 +1370,7 @@ ELFDEBUG( "*dest32=%8.8x\n", *dest32 );
 #endif
 	    break;
 	case R_PPC_REL32: /* 26 */
-	    dest32=(unsigned long *)(secp+rel->r_offset);
+	    dest32=(unsigned int *)(secp+rel->r_offset);
 #ifdef ELFDEBUG
 	    ELFDEBUG( "R_PPC_REL32\t" );
 	    ELFDEBUG( "secp=%x\t", secp );
